fix(libs-frontend-view): fix infinite recursion (#1989)
createAutocompleOptions function accepts and object with application state. When app contains an
input component (it will be present in state) - this caused an infinite recursion call which throws
error about max stack size reached. Some properties on an input causes this problem. Added
additional validation: if the object in state is a DOM node- do not process it recursively