fix(apps): #1553 fix commit pattern
"yarn commit" command used to generate commit message that does not pass validation. As a result
command failed and no commit was created. There is no option which may fix this error, so it was
agreed to change the commit format. Now issue number should be at the beginning of the title. Now
"yarn commit" will create the message which passes validation