fix(libs-frontend-view): apply default values from Form schema (#2059)
TypedValue form fields (ElementType, RenderPropsType, ReactNodeType, UnionType, ActionType) are
represented as an object with two properties: value (with actual id of element) and type (with id of
on of 5 TypedValue types). On the form only the field for "value" is rendered, "type" field is
hidden and there is no way to select it manually. And when on Props form a value is set to
TypedValue field - only "value" field is populated in form data which causes script error on attempt
to render component with this incomplete data. This commit adds mechanism which applies default
values that are defined in schema which fixes the issue. Once any TypedValue is set - "type"
property is populated with valid default value from schema.